pyproject.toml
[project]
name = "nec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.6.2",
    "crispy-bootstrap5>=2026.3",
    "django>=6.0.3",
    "django-auditlog>=3.4.1",
    "django-crispy-forms>=2.6",
    "django-fsm>=3.0.1",
    "django-htmx>=1.27.0",
    "django-money>=3.6.0",
    "django-reversion>=6.1.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "python-decouple>=3.8",
    "stripe>=14.4.1",
    "whitenoise>=6.12.0",
]
